+++ packages/composites/historical-average/test/unit/adapter.test.ts
@@ -129,19 +129,19 @@ describe('execute', () => {
 describe('getFromToDates', () => {
   it('should ignore days when both to- and from-dates are provided', () => {
     const { fromDate, toDate } = getFromToDates('2021-11-01', '2021-11-02', 100)
-    expect(fromDate.toISOString()).toBe('2021-11-01T00:00:00.000Z')
-    expect(toDate.toISOString()).toBe('2021-11-02T00:00:00.000Z')
+    expect(fromDate.toDateString()).toBe(new Date('2021-11-01').toDateString())
+    expect(toDate.toDateString()).toBe(new Date('2021-11-02').toDateString())
   })
 
   it('should add days to fromDate when toDate is omitted', () => {
     const { fromDate, toDate } = getFromToDates('2021-11-01', undefined, 7)
-    expect(fromDate.toISOString()).toBe('2021-11-01T00:00:00.000Z')
-    expect(toDate.toISOString()).toBe('2021-11-08T00:00:00.000Z')
+    expect(fromDate.toDateString()).toBe(new Date('2021-11-01').toDateString())
+    expect(toDate.toDateString()).toBe(new Date('2021-11-08').toDateString())
   })
 
   it('should subtract days from toDate when fromDate is omitted', () => {
     const { fromDate, toDate } = getFromToDates(undefined, '2021-11-08', 7)
-    expect(fromDate.toISOString()).toBe('2021-11-01T00:00:00.000Z')
-    expect(toDate.toISOString()).toBe('2021-11-08T00:00:00.000Z')
+    expect(fromDate.toDateString()).toBe(new Date('2021-11-01').toDateString())
+    expect(toDate.toDateString()).toBe(new Date('2021-11-08').toDateString())
   })
 })