+++ packages/examples/source/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/example-source-adapter
 
+## 1.0.19
+
+### Patch Changes
+
+- 8fdada79e: src export changed
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.0.18
 
 ### Patch Changes