+++ packages/sources/cryptoapis-v2/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/cryptoapis-v2-adapter
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/ea-factories@1.0.19
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.1.3
 
 ### Patch Changes